result mo

+++ practice/src/main/java/com/snort/practice/Repository/UserQuestionRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface UserQuestionRepository extends JpaRepository<UserQuestion,Long> {
-    List<UserQuestion> findByUserId(Long userId);
+    List<UserQuestion> findByUserId(int userId);
 //Page<UserQuestion> findByUserId(Long userId, Pageable pageable);
 }
